undo common compiler worker
